Added DAE solvers
