+++ lib/OrdinaryDiffEqDAE/Project.toml
@@ -1,0 +1,20 @@
+name = "OrdinaryDiffEqDAE"
+uuid = "7d2d7016-1c68-401d-834f-6be2c6dc7d88"
+authors = ["ParamThakkar123 <paramthakkar864@gmail.com>"]
+version = "1.0.0"
+
+[deps]
+OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
+TruncatedStacktraces = "781d530d-4396-4725-bb49-402e4bee1e77"
+
+[extras]
+DiffEqDevTools = "f3b72e0c-5b89-59e1-b016-84e28bfd966d"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[compat]
+julia = "1.10"
+
+[targets]
+test = ["DiffEqDevTools", "Random", "SafeTestsets", "Test"]
